Protect against incorrectly decrypted files being marked as [un]safe
This was an attack that was possible due to the fact that the MXC
was assumed to be the only input to determine the AV result. In
practice, the decryption key is also input so an attacker could
in theory construct a key that would mark a file as safe/unsafe
for all other users.

The prevention is simply to hash the entire input, including keys
and IV used in decryption of the media, and pin results against
the hash, which is a secret given back to the client.

The client then downloads the media via the media repo and includes
the secret in the request to download, giving the repo enough to
request the /scan_report from matrix-content-scanner.