Separate getReport from generateReport
This is mostly for easier testing of getReport; we need to test
the scenario where the file hasn't been scanned before without
causing a scan to be done implicitly.

Moving the logic to generate a scan in its absence to handlers
means we can make sure getReport returns scanned: false when it
hasn't been scanned before.