Implement directDownload config option for downloading from media domain
This option effectively sets the baseUrl for a given request for media
with the domain replaced with the media's domain. This forces a direct
download from the media's content repo.

`baseUrl` is effectively unused with this option enabled, but this is
probably better than enabling `directDownload` when `baseUrl` is not
specified. The option should only be used in the minority of cases
where the content scanner has restricted access to a set of trusted
content repositories so as to prevent downloading and scanning of
content from *any* domain.