Implement middleware to decipher encrypted POST requests
Add a generic middleware for POST endpoints that allows them to
accept requests that contain enciphered JSON strings.

The request body has the following structure:

```json
{
    "ephemeral": "[base64-encoded ephemeral public key]",
    "mac": "[base64-encodeed MAC]",
    "ciphertext": "[base64-encoded ciphertext of the JSON body]"
}
```

The above can be calculated with matrix-olm `PkEncryption`. The
public key of the server is required to do this and can be found
by doing a GET request to the `.../media_proxy/unstable/public_key`
endpoint.