Implement persistence of encrypted_body decryption key
via pickling of the Olm.PkDecryption object.

This required new configuration for the pickling secret (`pickleKey`)
and the path to the pickled decryption key itself (`picklePath`).

This also required making app creation asynchronous because the
pickle file is loaded when the path is provided. In-turn this required
all handler tests to be updated to be async and to each create an app
every time a test is run.