Fix unit tests
The CI was failing when the script is 'exit X' because the content scanner tries to invoke it with the file to scan as its last parameter.
This fixes it by adding the character '#' to the script passed through the config in these cases, so that any further parameter is considered a comment and isn't interpreted by bash.