Additionally check file mimetype after decryption
Before this change, we would only check the mimetype of an encrypted file based
on the information provided with the file by the client. This could allow clients
to upload executables but state that the mimetype was an innocuous image/jpeg.

This change now additionally checks the mimetype of encrypted files based on the
file contents (after it has been decrypted) - just as we do for non-encrypted files.