Update olm to v3.x (#42)
We were using Olm v2 still. This PR updates the dependency of Olm to the latest version (v3.2.3).

As this is bumping a major version of Olm, there was a breaking change. Namely that the [olm library must now be initialised](https://gitlab.matrix.org/matrix-org/olm/-/blob/master/CHANGELOG.rst). I opted to do so in `createApp`, and assigned the initialised `Olm` object to NodeJS's `global` object, which is accessible across the current process.

The tests were a bit messier. I couldn't easily figure out a nice place to initialise Olm once for all tests. So instead, we initialise in each test file where necessary.