Fix decrypted file mimetype checks
The decrypted path only fails the file on unknown mimetype if we're actually configured with a list of mimetypes. Add the missing `if` to look at that config.

Also add a bit of logging when files are skipped due to mimetype and make some let's const, since we don't redeclare them.